refactor(desktop): extract markdown file card
Move the generic attachment file card into the markdown submodule without changing its implementation.
